BUGFIX: The ASM setup of the main function parameters was wrong. The solution is actually far simpler.  The format is already correct from the kernel.  All that is needed is to get 3 function parameters off the stack.  I fixed it in "caller.s".
I changed the stack pointer alignment to be more robust.  While, I
do not think this is necessary, it now has the logic to fix the
stack pointer to a 16 byte alignment even if the stack pointer is off
by some non-zero value which fits inside the bitmask 0b111.  Or, in
other words, if the stack pointer is not even aligned 8 byte bounds,
The new code should fix it.  However, I don't think that is possible
in 64-bit mode.  It is there now, just in case, and doesn't add any
complexity or code to implement.